requirements_local.txt
asgiref==3.5.2
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.6
